Add files via upload
Initial data set that the analysis and code is based off. Contains all the holdings and weights of the ETF which we wish to model our optimised portfolio on. I chose Blackrocks ASX200 ETF which currently trades on the Australian Securities Exchange under the ticker 'IOZ'. More about the ETF and Blackrock can be found here: https://www.blackrock.com/au/individual/products/251852/ishares-core-s-and-p-asx-200-etf